+++ lib/v2/bdys/maintainer.js
@@ -1,0 +1,3 @@
+module.exports = {
+    '/:caty?/:type?/:area?/:year?/:order?': ['nczitzk'],
+};
