+++ lib/v2/bdys/router.js
@@ -1,0 +1,3 @@
+module.exports = function (router) {
+    router.get('/:caty?/:type?/:area?/:year?/:order?', require('./index'));
+};
