+++ lib/v2/bdys/templates/desc.art
@@ -1,0 +1,21 @@
+{{ if info }}
+{{@ info }}<br>
+{{ /if }}
+
+{{ if synopsis }}
+{{@ synopsis }}<br>
+{{ /if}}
+
+{{ if links }}
+<div><b>下载地址：</b>
+    {{ each links link }}
+    <div class="item"><div class="content">{{ link.downloadCategory.name }}: <a href="{{ link.url }}">{{ link.url }}</a></div></div>
+    {{ /each }}
+</div>
+{{ /if }}
+
+{{ if torrents }}
+<div><b>种子列表：</b>
+{{@ torrents }}
+</div>
+{{ /if }}
