--- lib/v2/mp4er/maintainer.js
@@ -1,3 +1,0 @@
-module.exports = {
-    '/:type?/:caty?/:area?/:year?/:order?': ['nczitzk'],
-};
