--- lib/v2/mp4er/radar.js
@@ -1,13 +1,0 @@
-module.exports = {
-    '52bdys.com': {
-        _name: '哔嘀影视',
-        '.': [
-            {
-                title: '首页',
-                docs: 'https://docs.rsshub.app/multimedia.html#bi-di-ying-shi',
-                source: ['/'],
-                target: '/mp4er/:type?/:caty?/:area?/:year?/:order?',
-            },
-        ],
-    },
-};
