--- lib/v2/mp4er/router.js
@@ -1,3 +1,0 @@
-module.exports = function (router) {
-    router.get('/:type?/:caty?/:area?/:year?/:order?', require('./index'));
-};
